Fixed calendarView and added month labels and weekday labels.
