--- iOS Dev Calendar/Calendar/DayTypeUIColorExtension.swift
@@ -1,36 +1,0 @@
-//
-//  DayTypeUIColorExtension.swift
-//  iOS Dev Calendar
-//
-//  Created by Wesley Keetch on 4/16/25.
-//
-
-
-//import SwiftUI
-//
-//extension DayType {
-//  var uiColor: UIColor {
-//    switch self {
-//    case .sf:
-//      return UIColor.systemBlue
-//    case .tp:
-//      return UIColor.systemGreen
-//    case .nd:
-//      return UIColor.systemYellow
-//    case .st:
-//      return UIColor.systemOrange
-//    case .tt:
-//      return UIColor.systemRed
-//    case .fa:
-//      return UIColor.systemPink
-//    case .pc:
-//      return UIColor.systemPurple
-//    case .gc:
-//      return UIColor.systemMint
-//    case .holiday:
-//      return UIColor.systemGray3
-//    case .other:
-//      return UIColor.systemGray
-//    }
-//  }
-//}
