--- iOS Dev Calendar/Calendar/Month Labels/MonthLabel.Center.swift
@@ -1,25 +1,0 @@
-//
-//  MonthLabel.Center.swift of CalendarView Demo
-//
-//  Created by Alina Petrovska on 01.12.2023.
-//    - Mail: alina.petrovskaya@mijick.com
-//    - GitHub: https://github.com/Mijick
-//
-//  Copyright ©2023 Mijick. Licensed under MIT License.
-//
-//import Foundation
-//import MijickCalendarView
-//import SwiftUI
-//
-//extension ML { struct Center: MonthLabel {
-//  let month: Date
-//
-//  func createContent() -> AnyView {
-//    Text(getString(format: "MMMM yyyy")
-//      .uppercased())
-//    .font(.bold(18))
-//    .foregroundStyle(.primary)
-//    .frame(maxWidth: .infinity, alignment: .center)
-//    .erased()
-//  }
-//}}
